ArticleReply styling and layout
